Introduces Event::Group
To DRY AddedUserToGroup / RemovedUserFromGroup
