+++ src/api/app/models/event/added_user_to_group.rb
@@ -1,30 +1,12 @@
 module Event
-  class AddedUserToGroup < Base
+  class AddedUserToGroup < Group
     self.description = 'Added member to group'
-    payload_keys :group, :member, :who
-
-    receiver_roles :member
-
     self.notification_explanation = 'Receive notifications when you are added to a group.'
 
     def subject
       return "You were added to the group '#{payload['group']}'" unless payload['who']
 
       "'#{payload['who']}' added you to the group '#{payload['group']}'"
-    end
-
-    def members
-      [User.find_by(login: payload['member'])]
-    end
-
-    def originator
-      payload_address('who')
-    end
-
-    def parameters_for_notification
-      super.merge({ notifiable_type: 'Group',
-                    notifiable_id: Group.find_by(title: payload['group']).id,
-                    type: 'NotificationGroup' })
     end
   end
 end
