+++ src/api/app/models/event/removed_user_from_group.rb
@@ -1,30 +1,12 @@
 module Event
-  class RemovedUserFromGroup < Base
+  class RemovedUserFromGroup < Group
     self.description = 'Removed member from group'
-    payload_keys :group, :member, :who
-
-    receiver_roles :member
-
     self.notification_explanation = 'Receive notifications when you are removed from a group.'
 
     def subject
       return "You were removed from the group '#{payload['group']}'" unless payload['who']
 
       "'#{payload['who']}' removed you from the group '#{payload['group']}'"
-    end
-
-    def members
-      [User.find_by(login: payload['member'])]
-    end
-
-    def originator
-      payload_address('who')
-    end
-
-    def parameters_for_notification
-      super.merge({ notifiable_type: 'Group',
-                    notifiable_id: Group.find_by(title: payload['group']).id,
-                    type: 'NotificationGroup' })
     end
   end
 end
